Changes for aggregation pipeline lookup
