+++ db/aggregate.go
@@ -3,6 +3,7 @@ package mongik
 import (
 	"context"
 	"fmt"
+	"strings"
 
 	"github.com/FrosTiK-SD/mongik/constants"
 	mongik "github.com/FrosTiK-SD/mongik/models"
@@ -14,6 +15,20 @@ func Aggregate[Result any](mongikClient *mongik.Mongik, db string, collectionNam
 	var resultBytes []byte
 	var result []Result
 	var resultInterface []map[string]interface{}
+
+	// Parsing lookup collection from pipeline
+	var lookupCollection string = " "
+	pipe := fmt.Sprintf("%v", pipeline)
+	pipeSplit := strings.Split(pipe, "$lookup")
+	if len(pipeSplit) > 1 {
+		pipeSplit2 := strings.Split(pipeSplit[1], " ")
+		for _, tag := range pipeSplit2 {
+			if strings.Contains(tag, "from:") {
+				res := strings.Split(tag, "from:")
+				lookupCollection = res[1]
+			}
+		}
+	}
 
 	// First Check if it is present in the cache
 	if !noCache {
@@ -37,7 +52,7 @@ func Aggregate[Result any](mongikClient *mongik.Mongik, db string, collectionNam
 
 	// Set to cache
 	resultBytes, _ = json.Marshal(result)
-	if err := DBCacheSet(mongikClient.CacheClient, key, resultBytes); err == nil {
+	if err := DBCacheSet(mongikClient.CacheClient, key, resultBytes, lookupCollection); err == nil {
 		fmt.Println("Successfully set DB call in cache with key ", key)
 	}
